VMS.Collector
