+++ python/lsst/ts/cRIOpy/VMS/__init__.py
@@ -1,6 +1,7 @@
 from .Bars import StatusBar, ToolBar
 from .BoxChartWidget import BoxChartWidget
 from .Cache import Cache
+from .Collector import Collector, VMS_DEVICES
 from .ChartView import ChartView
 from .CSCPSDWidget import CSCPSDWidget
 from .DisplacementWidget import DisplacementWidget
